ditto_connection/requirements.txt
requests==2.23.0
influxdb>=5.3.0
sseclient-py
